shortenExternalImage util function
