+++ apps/src/codebridge/FilePreview/HTMLPreview.tsx
@@ -3,7 +3,7 @@ import {CodebridgeEmptyState} from '@codebridge/components/CodebridgeEmptyState'
 import {setCspViolations} from '@codebridge/redux/workspaceRedux';
 import classNames from 'classnames';
 import {isEqual} from 'lodash';
-import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import React, {useEffect, useMemo, useRef, useState} from 'react';
 
 import codebridgeI18n from '@cdo/apps/codebridge/locale';
 import useLifecycleNotifier from '@cdo/apps/lab2/hooks/useLifecycleNotifier';
@@ -22,6 +22,7 @@ import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import experiments from '@cdo/apps/util/experiments';
 import {useAppSelector, useAppDispatch} from '@cdo/apps/util/reduxHooks';
 
+import {shortenExternalURL} from '../utils';
 import {filterSourceForPreview} from '../utils/filterSourceForPreview';
 
 import {
@@ -134,36 +135,6 @@ export const HTMLPreview: React.FC = () => {
   const canNavigateBack = navigationHistoryIndex > 0;
   const canNavigateForward =
     navigationHistoryIndex < navigationHistory.length - 1;
-
-  const shortenURL = useCallback((url: string) => {
-    try {
-      const urlObj = new URL(url);
-      const hostname = urlObj.hostname;
-      const pathname = urlObj.pathname;
-      if (pathname === '/' || pathname === '') {
-        return hostname; // No path, just return hostname
-      }
-      // Get first directory after hostname
-      const pathParts = pathname.split('/').filter(part => part !== '');
-      if (pathParts.length === 0) {
-        return hostname;
-      }
-      // Return hostname plus path to next forward slash
-      return `${hostname}/${pathParts[0]}`;
-    } catch {
-      // If URL parsing fails, try to extract manually
-      const withoutProtocol = url.replace(/^https?:\/\//, '');
-      const parts = withoutProtocol.split('/').filter(part => part !== '');
-      if (parts.length === 0) {
-        return url; // Fallback to original
-      }
-      if (parts.length === 1) {
-        return parts[0]; // Just hostname
-      }
-      // Return hostname plus first directory
-      return `${parts[0]}/${parts[1]}`;
-    }
-  }, []);
 
   useEffect(() => {
     if (aiFilePathToPreview) {
@@ -336,7 +307,7 @@ export const HTMLPreview: React.FC = () => {
           blockedURI &&
           !cspViolations.some(v => v.blockedURI === blockedURI)
         ) {
-          const shortURL = shortenURL(blockedURI) + '...';
+          const shortURL = shortenExternalURL(blockedURI) + '...';
           dispatch(
             setCspViolations([
               ...cspViolations,
@@ -357,7 +328,6 @@ export const HTMLPreview: React.FC = () => {
     debouncedSource,
     dispatch,
     cspViolations,
-    shortenURL,
   ]);
 
   useEffect(() => {
